fix: Refactor food db

+++ backend/src/main/java/com/hankki/common/vector/VectorAggregator.java
@@ -7,15 +7,15 @@ public class VectorAggregator {
     /**
      * 주어진 float 벡터 리스트를 평균 벡터로 그룹화
      */
-    public static float[] average(List<float[]> vectors) {
+    public static double[] average(List<double[]> vectors) {
         if (vectors == null || vectors.isEmpty()) {
             throw new IllegalArgumentException("벡터 리스트가 비어있습니다.");
         }
 
         int dim = vectors.get(0).length;
-        float[] avg = new float[dim];
+        double[] avg = new double[dim];
 
-        for (float[] vec : vectors) {
+        for (double[] vec : vectors) {
             for (int i = 0; i < dim; i++) {
                 avg[i] += vec[i];
             }
